refactor: fixed sonarcloud code smells
