+++ driver/src/data/index.ts
@@ -3,5 +3,5 @@ export { default as DeepInsertService } from './deepInsertService';
 export { DeepInsertResponse } from './deepInsertResponse';
 export { Record } from './record';
 export { TestRecord } from './testRecord';
-export { default as Preprocessor } from './preprocesser';
+export { default as Preprocessor } from './preprocessor';
 export { default as FakerPreprocessor } from './fakerPreprocessor';
